+++ addon/index.js
@@ -1,0 +1,2 @@
+import { useEmberService } from './hooks';
+export { useEmberService };