+++ tests/dummy/app/services/test.js
@@ -1,0 +1,7 @@
+import Service from '@ember/service';
+
+export default class TestService extends Service {
+  getText() {
+    return 'Hello from Test Service';
+  }
+}